unified bundleUtils in client and server removed duplicated utils in server
